refactor(styles): migrates to styled-components
fixes https://github.com/frontity/saturn-theme/issues/113
