+++ src/shared/components/HtmlToReactConverter/index.js
@@ -5,7 +5,7 @@ import { inject } from 'mobx-react';
 import { parse } from 'himalaya';
 import he from 'he';
 import { flow, camelCase, capitalize } from 'lodash';
-import { withTheme } from 'emotion-theming';
+import { withTheme } from 'styled-components';
 
 import injectSlots from './injectSlots';
 import { filter } from './filter';
